requirements.txt
torch>=1.12.0
torchvision>=0.13.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
pandas>=1.3.0
PyYAML>=6.0
tensorboard>=2.8.0
numpy>=1.21.0
Pillow>=8.3.0
datasets>=2.0.0
huggingface-hub>=0.10.0
transformers>=4.20.0